+++ pennylane/plugins/default_qubit.py
@@ -370,6 +370,11 @@ class DefaultQubit(Device):
 
         return ev
 
+    def var(self, expectation, wires, par):
+        # measurement/expectation value <psi|A|psi>
+        A = self._get_operator_matrix(expectation, par)
+        return self.ev(A@A, wires) - self.ev(A, wires)**2
+
     def _get_operator_matrix(self, operation, par):
         """Get the operator matrix for a given operation or observable.