+++ pennylane/qnode.py
@@ -374,8 +374,8 @@ class QNode:
             raise QuantumFunctionError("All measured observables must be returned in the "
                                        "order they are measured.")
 
-        self.ev = res  #: tuple[Observable]: returned observables
-        self.ops = self.queue + list(self.ev)  #: list[Operation]: combined list of circuit operations
+        self.ev = list(res)  #: list[Observable]: returned observables
+        self.ops = self.queue + self.ev  #: list[Operation]: combined list of circuit operations
 
         # classify the circuit contents
         temp = [isinstance(op, pennylane.operation.CV) for op in self.ops if not isinstance(op, pennylane.ops.Identity)]
@@ -485,6 +485,10 @@ class QNode:
                         if x.ev_order is None:
                             return 'F'
                         if x.ev_order == 2:
+                            if x.return_type == 'variance':
+                                # second order observables don't support
+                                # analytic diff of variances
+                                return 'F'
                             op.grad_method = 'A2'  # bit of a hack
                     return 'A'
 
@@ -703,6 +707,8 @@ class QNode:
             else:
                 y0 = None
 
+        variances = any(e.return_type == 'variance' for e in self.ev)
+
         # compute the partial derivative w.r.t. each parameter using the proper method
         grad = np.zeros((self.output_dim, len(which)), dtype=float)
 
@@ -713,7 +719,10 @@ class QNode:
 
             par_method = method[k]
             if par_method == 'A':
-                grad[:, i] = self._pd_analytic(flat_params, k, **kwargs)
+                if variances:
+                    grad[:, i] = self._pd_analytic_var(flat_params, k, **kwargs)
+                else:
+                    grad[:, i] = self._pd_analytic(flat_params, k, **kwargs)
             elif par_method == 'F':
                 grad[:, i] = self._pd_finite_diff(flat_params, k, h, order, y0, **kwargs)
             else:
@@ -863,6 +872,116 @@ class QNode:
             op.params[p_idx] = orig
 
         return pd
+
+    def _pd_analytic_var(self, param_values, param_idx, **kwargs):
+        """Partial derivative of variances of observables using the analytic method.
+
+        Args:
+            param_values (array[float]): point in free parameter space at which
+                to evaluate the partial derivative
+            param_idx (int): return the partial derivative with respect to this
+                free parameter
+
+        Returns:
+            float: partial derivative of the node.
+        """
+        old_ev = copy.deepcopy(self.ev)
+
+        # boolean mask: elements are True where the
+        # return type is a variance, False for expectations
+        where_var = [e.return_type == "variance" for e in self.ev]
+
+        for i, e in enumerate(self.ev):
+            # iterate through all observables
+            # here, i is the index of the observable
+            # and e is the observable
+
+            if e.return_type != 'variance':
+                # if the expectation value is not a variance
+                # continue on to the next loop iteration
+                continue
+
+            # temporarily convert return type to expectation
+            self.ev[i].return_type = 'expectation'
+
+            # analytic derivative of <A^2>
+            # For involutory observables (A^2 = I),
+            # then d<I>/dp = 0
+            pdA2 = 0
+
+            if self.type == 'qubit':
+                if e.__class__.__name__ == 'Hermitian':
+                    # since arbitrary Hermitian observables
+                    # are not guaranteed to be involutory, need to take them into
+                    # account separately to calculate d<A^2>/dp
+
+                    A = e.params[0]  # Hermitian matrix
+                    w = e.wires
+
+                    if not np.allclose(A @ A, np.identity(A.shape[0])):
+                        # make a copy of the original variance
+                        old = copy.deepcopy(e)
+
+                        # replace the Hermitian variance with <A^2> expectation
+                        self.ev[i] = pennylane.expval(pennylane.ops.Hermitian(A @ A, w, do_queue=False))
+
+                        # calculate the analytic derivative of <A^2>
+                        pdA2 = np.asarray(self._pd_analytic(param_values, param_idx, **kwargs))
+
+                        # restore the original Hermitian variance
+                        self.ev[i] = old
+
+            elif self.type == 'CV':
+                # need to calculate d<A^2>/dp
+                # make a copy of the original variance
+                old = copy.deepcopy(e)
+                w = old.wires
+
+                # get the heisenberg representation
+                # This will be a real 1D vector representing the
+                # first order observable in the basis [I, x, p]
+                A = e._heisenberg_rep(old.parameters) # pylint: disable=protected-access
+
+                # make this a row vector by adding an extra dimension
+                A = np.expand_dims(A, axis=0)
+
+                # take the outer product of the heisenberg representation
+                # with itself, to get a square symmetric matrix representing
+                # the square of the observable
+                A = np.kron(A, A.T)
+
+                # replace the first order observable var(A) with <A^2>
+                # in the return queue
+                self.ev[i] = pennylane.expval(pennylane.ops.PolyXP(A, w, do_queue=False))
+
+                # calculate the analytic derivative of <A^2>
+                pdA2 = np.asarray(self._pd_analytic(param_values, param_idx, force_order2=True, **kwargs))
+
+                # restore the original observable
+                self.ev[i] = old
+
+        # save original cache setting
+        cache = self.cache
+        # Make sure caching is on. If it is not on,
+        # the circuit will be reconstructed when self.evaluate is
+        # called, overwriting the temporary change we made to
+        # self.ev, where we set the return_type of every observable
+        # to 'expectation'.
+        self.cache = True
+
+        # evaluate circuit value at original parameters
+        evA = np.asarray(self.evaluate(param_values, **kwargs))
+        # evaluate circuit gradient assuming all outputs are expectations
+        pdA = self._pd_analytic(param_values, param_idx, **kwargs)
+
+        # restore original return queue
+        self.ev = old_ev
+        # restore original caching setting
+        self.cache = cache
+
+        # return the variance shift rule where where_var==True,
+        # otherwise return the expectation parameter shift rule
+        return np.where(where_var, pdA2-2*evA*pdA, pdA)
 
     def to_torch(self):
         """Convert the standard PennyLane QNode into a :func:`~.TorchQNode`.