+++ tests/test_default_qubit.py
@@ -423,6 +423,49 @@ class TestDefaultQubitDevice(BaseTest):
             with pytest.warns(RuntimeWarning, match='Nonvanishing imaginary part'):
                 self.dev.ev(H + 1j, [0])
 
+    def test_var_pauliz(self):
+        """Test that variance of PauliZ is the same as I-<Z>^2"""
+        self.logTestName()
+        self.dev.reset()
+
+        phi = 0.543
+        theta = 0.6543
+        self.dev.apply('RX', wires=[0], par=[phi])
+        self.dev.apply('RY', wires=[0], par=[theta])
+
+        var = self.dev.var('PauliZ', [0], [])
+        mean = self.dev.expval('PauliZ', [0], [])
+
+        self.assertAlmostEqual(var, 1-mean**2, delta=self.tol)
+
+    def test_var_pauliz_rotated_state(self):
+        """test correct variance for <Z> of a rotated state"""
+        self.logTestName()
+        self.dev.reset()
+
+        phi = 0.543
+        theta = 0.6543
+        self.dev.apply('RX', wires=[0], par=[phi])
+        self.dev.apply('RY', wires=[0], par=[theta])
+        var = self.dev.var('PauliZ', [0], [])
+        expected = 0.25*(3-np.cos(2*theta)-2*np.cos(theta)**2*np.cos(2*phi))
+        self.assertAlmostEqual(var, expected, delta=self.tol)
+
+    def test_var_hermitian_rotated_state(self):
+        """test correct variance for <H> of a rotated state"""
+        self.logTestName()
+        self.dev.reset()
+
+        phi = 0.543
+        theta = 0.6543
+
+        H = np.array([[4, -1+6j], [-1-6j, 2]])
+        self.dev.apply('RX', wires=[0], par=[phi])
+        self.dev.apply('RY', wires=[0], par=[theta])
+        var = self.dev.var('Hermitian', [0], [H])
+        expected = 0.5*(2*np.sin(2*theta)*np.cos(phi)**2+24*np.sin(phi)\
+                    *np.cos(phi)*(np.sin(theta)-np.cos(theta))+35*np.cos(2*phi)+39)
+        self.assertAlmostEqual(var, expected, delta=self.tol)
 
 class TestDefaultQubitIntegration(BaseTest):
     """Integration tests for default.qubit. This test ensures it integrates